requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
async-timeout==5.0.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.2.2
fastapi==0.112.0
greenlet==3.2.4
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
pillow==10.4.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyMySQL==1.1.2
PySocks==1.7.1
python-dotenv==1.2.1
python-jose==3.5.0
qiniu==7.17.0
redis==7.0.1
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
starlette==0.37.2
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
urllib3==2.2.2
uvicorn==0.38.0
websocket-client==1.8.0
wget==3.2
wsproto==1.2.0
typing_extensions==4.12.2
rq==1.15.1
pyarmor==8.5.7
